Create base class for manual sync pages
