+++ cypress/pageobjects/ManualAggregateSyncPageObject.js
@@ -1,56 +1,13 @@
 import { dataTest } from "../support/utils";
+import ManualSyncPageObject from "./common/ManualSyncPageObject";
 
-class ManualAggregateSyncPageObject {
+class ManualAggregateSyncPageObject extends ManualSyncPageObject {
     constructor(cy) {
-        this.cy = cy;
-    }
-
-    get title() {
-        return this.cy.get(dataTest("page-header-title"));
-    }
-
-    get error() {
-        return this.cy.get("#client-snackbar");
-    }
-
-    get syncResults() {
-        return this.cy.get('[data-test="Typography-synchronization-results-row-0"]');
-    }
-
-    get syncButton() {
-        return this.cy.get('[data-test="Button-aggregated-data-synchronization-save"]');
+        super(cy, "aggregated-data");
     }
 
     open() {
-        this.cy.login("admin");
-        this.cy.visit("/#/sync/aggregated");
-        this.cy.get(dataTest("headerbar-title")).contains("MetaData Synchronization");
-        return this;
-    }
-
-    search(text) {
-        this.cy.get('[data-test="search"] > div > [aria-invalid="false"]').type(text);
-        return this;
-    }
-
-    selectRow(text) {
-        this.cy
-            .get(dataTest("TableCell-data-table-row-0-column-displayname"))
-            .contains(text)
-            .click();
-        return this;
-    }
-
-    openSyncDialog(text) {
-        this.cy
-            .get('[data-test="objects-table-action-button"] > :nth-child(1) > [focusable="false"]')
-            .click();
-
-        return this;
-    }
-
-    next() {
-        this.cy.get('[data-test="Button-next-→"]').click();
+        super.open("/#/sync/aggregated");
         return this;
     }
 
@@ -79,22 +36,6 @@ class ManualAggregateSyncPageObject {
         return this;
     }
 
-    selectReceiverInstance(instance) {
-        // Move to instance selection
-        this.cy
-            .get(dataTest("DialogContent-aggregated-data-synchronization"))
-            .contains("Instance Selection")
-            .click();
-        this.cy.waitForStep("Instance Selection");
-
-        // Select receiver instance (multi-selector does not work fine with cypress?)
-        this.cy.selectInMultiSelector(
-            dataTest("DialogContent-aggregated-data-synchronization"),
-            instance
-        );
-        return this;
-    }
-
     synchronize() {
         this.cy
             .route({
@@ -105,11 +46,6 @@ class ManualAggregateSyncPageObject {
 
         this.syncButton.click();
         this.cy.wait("@postDataValueSets");
-        return this;
-    }
-
-    closeSyncResultsDialog() {
-        this.cy.get('[data-test="Button-synchronization-results-save"]').click();
         return this;
     }
 }
